Move invoke_training.training.shared.data to new training2 directory.
