+++ src/invoke_training/training/textual_inversion/textual_inversion_sd.py
@@ -9,19 +9,12 @@ import torch
 import torch.utils.checkpoint
 from accelerate import Accelerator
 from accelerate.utils import set_seed
-from diffusers import (
-    AutoencoderKL,
-    DDPMScheduler,
-    StableDiffusionPipeline,
-    UNet2DConditionModel,
-)
+from diffusers import AutoencoderKL, DDPMScheduler, StableDiffusionPipeline, UNet2DConditionModel
 from diffusers.optimization import get_scheduler
 from tqdm.auto import tqdm
 from transformers import CLIPTextModel, CLIPTokenizer
 
-from invoke_training.training.config.textual_inversion_config import (
-    TextualInversionConfig,
-)
+from invoke_training.training.config.textual_inversion_config import TextualInversionConfig
 from invoke_training.training.finetune_lora.finetune_lora_sd import (
     cache_vae_outputs,
     generate_validation_images,
@@ -33,17 +26,11 @@ from invoke_training.training.shared.accelerator_utils import (
     initialize_logging,
 )
 from invoke_training.training.shared.checkpoint_tracker import CheckpointTracker
-from invoke_training.training.shared.data.data_loaders.textual_inversion_sd_dataloader import (
-    build_textual_inversion_sd_dataloader,
-)
-from invoke_training.training.shared.model_loading_utils import (
-    PipelineVersionEnum,
-    load_pipeline,
-)
+from invoke_training.training.shared.model_loading_utils import PipelineVersionEnum, load_pipeline
 from invoke_training.training.shared.optimizer_utils import initialize_optimizer
-from invoke_training.training.shared.serialization import (
-    load_state_dict,
-    save_state_dict,
+from invoke_training.training.shared.serialization import load_state_dict, save_state_dict
+from invoke_training.training2.shared.data.data_loaders.textual_inversion_sd_dataloader import (
+    build_textual_inversion_sd_dataloader,
 )
