+++ src/invoke_training/training2/shared/data/ARCHITECTURE.md
