+++ src/invoke_training/training2/shared/data/datasets/hf_dir_image_caption_dataset.py
