+++ src/invoke_training/training2/shared/data/datasets/image_dir_dataset.py
