+++ src/invoke_training/training2/shared/data/transforms/drop_field_transform.py
