+++ src/invoke_training/training2/shared/data/transforms/sd_image_transform.py
