+++ src/invoke_training/training2/shared/data/transforms/sdxl_image_transform.py
