+++ src/invoke_training/training2/shared/data/transforms/shuffle_caption_transform.py
