+++ src/invoke_training/training2/shared/data/transforms/tensor_disk_cache.py
