+++ src/invoke_training/training2/shared/data/transforms/textual_inversion_caption_transform.py
