+++ tests/invoke_training/training/shared/data/datasets/test_transform_dataset.py
@@ -1,8 +1,6 @@
 import unittest.mock
 
-from invoke_training.training.shared.data.datasets.transform_dataset import (
-    TransformDataset,
-)
+from invoke_training.training2.shared.data.datasets.transform_dataset import TransformDataset
 
 
 def test_transform_dataset_len():
